#83 #84 #89 small update
add a comment button that can direct the user to a new page which should display all the detail of the pic, comment, likes, @someone, etc. 
There is a small bug where you have to press the button from the home page rather than the profile page. Since the home page displays the same thing as the profile page, they both have the button. Check the `app-routing.moudule.ts` file in order to fix this.