#81 #83 #84 should be working
Display comment, add comment and remove comment functions are working properly. There is a bug when refreshing the page, comment page will not receive data passing from gallery item. the URL of the comment page should be `xxx/commentPage/picId` rather than `xxx/commentPage`. Auto-refresh also needs to be added once submit and delete comments. During the test, do NOT refresh the page directly. Go to the home page and then click the comment button to go into the comment page. Therefore you can check the deleted and uploaded comments. If you the comment page, there will be a 404 error since the URL is not being passed from the gallery-item (the sibling of commentPage).